requirements.txt
matplotlib>=3.4.2
numpy>=1.22.0
Pillow>=8.3.2
scipy>=1.7.1
torch>=2.5
torchvision>=0.2